feat(cijoe): bump to v0.9.52
This upgrade comes with a lot of refactorisation, which therefore makes
this commit quite large.

- Remove `--workflow` argument, as this is no longer part of the cijoe
  cli. The workflow has been changed to a positional argument.
- Refactor config- and workflow files for new Qemu module steup.
- Add argparse arguments to scripts instead of step arguments and remove
  the step parameter

+++ cijoe/scripts/liburing_build.py
@@ -3,22 +3,26 @@
 Build liburing from source
 ==========================
 
-Step Args
----------
-
 Retargetable: True
 ------------------
 """
 import errno
+from argparse import ArgumentParser
 from pathlib import Path
 
 
-def main(args, cijoe, step):
+def add_args(parser: ArgumentParser):
+    parser.add_argument("--liburing_source", type=str, default=None)
+
+
+def main(args, cijoe):
     """Build liburing"""
 
-    liburing_source = step.get("with", {}).get(
-        "liburing_source", cijoe.getconf("liburing.repository.path", None)
+    liburing_source = args.liburing_source or cijoe.getconf(
+        "liburing.repository.path", None
     )
+    if not liburing_source:
+        return errno.EINVAL
 
     commands = [
         "unlink /usr/lib/liburing.so || true",
